Isolate runtime hash functions into its own module
