+++ opal/runtime/boot.js
@@ -151,6 +151,8 @@
     }
   }
 
+  Opal.raise = $raise;
+
   // Reuse the same object for performance/memory sake
   var prop_options = {
     value: undefined,
@@ -181,6 +183,8 @@
   Opal.defineProperty = Opal.prop;
 
   Opal.slice = $slice;
+  Opal.splice = $splice;
+  Opal.has_own = $has_own;
 
   // Helpers
   // -----
@@ -1806,27 +1810,6 @@
     return ancestors.indexOf(klass) !== -1;
   };
 
-  // Helpers for extracting kwsplats
-  // Used for: { **h }
-  Opal.to_hash = function(value) {
-    if (value.$$is_hash) {
-      return value;
-    }
-    else if (value['$respond_to?']('to_hash', true)) {
-      var hash = value.$to_hash();
-      if (hash.$$is_hash) {
-        return hash;
-      }
-      else {
-        $raise(Opal.TypeError, "Can't convert " + value.$$class +
-          " to Hash (" + value.$$class + "#to_hash gives " + hash.$$class + ")");
-      }
-    }
-    else {
-      $raise(Opal.TypeError, "no implicit conversion of " + value.$$class + " into Hash");
-    }
-  };
-
   // Helpers for implementing multiple assignment
   // Our code for extracting the values and assigning them only works if the
   // return value is a JS array.
@@ -1877,44 +1860,6 @@
     else {
       return [value];
     }
-  };
-
-  // Used for extracting keyword arguments from arguments passed to
-  // JS function.
-  //
-  // @param parameters [Array]
-  // @return [Hash] or undefined
-  //
-  Opal.extract_kwargs = function(parameters) {
-    var kwargs = parameters[parameters.length - 1];
-    if (kwargs != null && Opal.respond_to(kwargs, '$to_hash', true)) {
-      $splice(parameters, parameters.length - 1);
-      return kwargs;
-    }
-  };
-
-  // Used to get a list of rest keyword arguments. Method takes the given
-  // keyword args, i.e. the hash literal passed to the method containing all
-  // keyword arguments passed to method, as well as the used args which are
-  // the names of required and optional arguments defined. This method then
-  // just returns all key/value pairs which have not been used, in a new
-  // hash literal.
-  //
-  // @param given_args [Hash] all kwargs given to method
-  // @param used_args [Object<String: true>] all keys used as named kwargs
-  // @return [Hash]
-  //
-  Opal.kwrestargs = function(given_args, used_args) {
-    var map = new Map();
-
-    Opal.hash_each(given_args, false, function(key, value) {
-      if (!used_args[key]) {
-        Opal.hash_put(map, key, value);
-      }
-      return [false, false];
-    });
-
-    return map;
   };
 
   function apply_blockopts(block, blockopts) {
@@ -2305,273 +2250,6 @@
     Opal.defn(obj, id, body);
 
     return obj;
-  };
-
-
-  // Hashes
-  // ------
-
-
-  // Experiments have shown, that using new Map([[1,2]]) inline is rather slow
-  // compared to using new Map() in combination with .set(1,2), because the former
-  // creates a new Array for each pair and then discards it. Using .set though
-  // would increase the size of the generated code. So lets use a compromise and
-  // use a helper function, which allows the compiler to generate compact code
-  // and at the same time provides the performance improvement of using .set
-  // with a overall smaller overhead than creating arrays for each pair.
-  // For primitive keys:
-  Opal.hash_new = function () {
-    let h = new Map();
-    for (let i = 0; i < arguments.length; i += 2) {
-      h.set(arguments[i], arguments[i + 1]);
-    }
-    return h;
-  }
-
-  // The same as above, except for complex keys:
-  Opal.hash_new2 = function () {
-    let h = new Map();
-    for (let i = 0; i < arguments.length; i += 2) {
-      Opal.hash_put(h, arguments[i], arguments[i + 1]);
-    }
-    return h;
-  }
-
-  Opal.hash_init = function (_hash) {
-    console.warn("DEPRECATION: Opal.hash_init is deprecated and is now a no-op.");
-  }
-
-  Opal.hash_clone = function(from_hash, to_hash) {
-    to_hash.$$none = from_hash.$$none;
-    to_hash.$$proc = from_hash.$$proc;
-
-    return Opal.hash_each(from_hash, to_hash, function(key, value) {
-      Opal.hash_put(to_hash, key, value);
-      return [false, to_hash];
-    });
-  };
-
-  Opal.hash_put = function(hash, key, value) {
-    var type = typeof key;
-    if (type === "string" || type === "symbol" || type === "number" || type === "boolean" || type === "bigint") {
-      hash.set(key, value)
-    } else if (key.$$is_string) {
-      hash.set(key.valueOf(), value);
-    } else {
-      if (!hash.$$keys)
-        hash.$$keys = new Map();
-
-      var key_hash = key.$$is_string ? key.valueOf() : (hash.$$by_identity ? Opal.id(key) : key.$hash()),
-          keys = hash.$$keys;
-
-      if (!keys.has(key_hash)) {
-        keys.set(key_hash, [key]);
-        hash.set(key, value);
-        return;
-      }
-
-      var objects = keys.get(key_hash),
-          object;
-
-      for (var i=0; i<objects.length; i++) {
-        object = objects[i];
-        if (key === object || key['$eql?'](object)) {
-          hash.set(object, value);
-          return;
-        }
-      }
-
-      objects.push(key);
-      hash.set(key, value);
-    }
-  };
-
-  Opal.hash_get = function(hash, key) {
-    var type = typeof key;
-    if (type === "string" || type === "symbol" || type === "number" || type === "boolean" || type === "bigint") {
-      return hash.get(key)
-    } else if (hash.$$keys) {
-      var key_hash = key.$$is_string ? key.valueOf() : (hash.$$by_identity ? Opal.id(key) : key.$hash()),
-          objects = hash.$$keys.get(key_hash),
-          object;
-
-      if (objects !== undefined) {
-        for (var i=0; i<objects.length; i++) {
-          object = objects[i];
-          if (key === object || key['$eql?'](object))
-            return hash.get(object);
-        }
-      } else if (key.$$is_string) {
-        return hash.get(key_hash);
-      }
-    } else if (key.$$is_string) {
-      return hash.get(key.valueOf());
-    }
-  };
-
-  function $hash_delete_stage2(hash, key) {
-    var value = hash.get(key);
-    hash.delete(key);
-    return value;
-  }
-
-  Opal.hash_delete = function(hash, key) {
-    var type = typeof key
-    if (type === "string" || type === "symbol" || type === "number" || type === "boolean" || type === "bigint") {
-      return $hash_delete_stage2(hash, key);
-    } else if (hash.$$keys) {
-      var key_hash = key.$$is_string ? key.valueOf() : (hash.$$by_identity ? Opal.id(key) : key.$hash()),
-          objects = hash.$$keys.get(key_hash),
-          object;
-
-      if (objects !== undefined) {
-        for (var i=0; i<objects.length; i++) {
-          object = objects[i];
-          if (key === object || key['$eql?'](object)) {
-            objects.splice(i, 1);
-            if (objects.length === 0)
-              hash.$$keys.delete(key_hash);
-            return $hash_delete_stage2(hash, object);
-          }
-        }
-      } else if (key.$$is_string) {
-        return $hash_delete_stage2(hash, key_hash);
-      }
-    } else if (key.$$is_string) {
-      return $hash_delete_stage2(hash, key.valueOf());
-    }
-  };
-
-  Opal.hash_rehash = function(hash) {
-    var keys = hash.$$keys;
-
-    if (keys)
-      keys.clear();
-
-    Opal.hash_each(hash, false, function(key, value) {
-      var type = typeof key;
-      if (type === "string" || type === "symbol" || type === "number" || type === "boolean" || type === "bigint")
-        return [false, false]; // nothing to rehash
-
-      var key_hash = key.$$is_string ? key.valueOf() : (hash.$$by_identity ? Opal.id(key) : key.$hash());
-
-      if (!keys)
-        hash.$$keys = keys = new Map();
-
-      if (!keys.has(key_hash)) {
-        keys.set(key_hash, [key]);
-        return [false, false];
-      }
-
-      var objects = keys.get(key_hash),
-          objects_copy = (objects.length === 1) ? objects : $slice(objects),
-          object;
-
-      for (var i=0; i<objects_copy.length; i++) {
-        object = objects_copy[i];
-        if (key === object || key['$eql?'](object)) {
-          // got a duplicate, remove it
-          objects.splice(objects.indexOf(object), 1);
-          hash.delete(object);
-        }
-      }
-
-      objects.push(key);
-
-      return [false, false]
-    });
-
-    return hash;
-  };
-
-  Opal.hash = function () {
-    var arguments_length = arguments.length,
-      args,
-      hash,
-      i,
-      length,
-      key,
-      value;
-
-    if (arguments_length === 1 && arguments[0].$$is_hash) {
-      return arguments[0];
-    }
-
-    hash = new Map();
-
-    if (arguments_length === 1) {
-      args = arguments[0];
-
-      if (arguments[0].$$is_array) {
-        length = args.length;
-
-        for (i = 0; i < length; i++) {
-          if (args[i].length !== 2) {
-            $raise(Opal.ArgumentError, 'value not of length 2: ' + args[i].$inspect());
-          }
-
-          key = args[i][0];
-          value = args[i][1];
-
-          Opal.hash_put(hash, key, value);
-        }
-
-        return hash;
-      } else {
-        args = arguments[0];
-        for (key in args) {
-          if ($has_own(args, key)) {
-            value = args[key];
-
-            Opal.hash_put(hash, key, value);
-          }
-        }
-
-        return hash;
-      }
-    }
-
-    if (arguments_length % 2 !== 0) {
-      $raise(Opal.ArgumentError, 'odd number of arguments for Hash');
-    }
-
-    for (i = 0; i < arguments_length; i += 2) {
-      key = arguments[i];
-      value = arguments[i + 1];
-
-      Opal.hash_put(hash, key, value);
-    }
-
-    return hash;
-  };
-
-  // A faster Hash creator for hashes that just use symbols and
-  // strings as keys. The map and keys array can be constructed at
-  // compile time, so they are just added here by the constructor
-  // function.
-  //
-  Opal.hash2 = function(keys, smap) {
-    console.warn("DEPRECATION: `Opal.hash2` is deprecated and will be removed in Opal 2.0. Use $hash_new for primitive keys or $hash_new2 for complex keys instead.");
-
-    var hash = new Map();
-    for (var i = 0, max = keys.length; i < max; i++) {
-      hash.set(keys[i], smap[keys[i]]);
-    }
-    return hash;
-  };
-
-  Opal.hash_each = function (hash, dres, fun) {
-    // dres = default result, returned if hash is empty
-    // fun is called as fun(key, value) and must return a array with [break, result]
-    // if break is true, iteration stops and result is returned
-    // if break is false, iteration continues and eventually the last result is returned
-    var res;
-    for (var i = 0, entry, entries = Array.from(hash.entries()), l = entries.length; i < l; i++) {
-      entry = entries[i];
-      res = fun(entry[0], entry[1]);
-      if (res[0]) return res[1];
-    }
-    return res ? res[1] : dres;
   };
 
   // Create a new range instance with first and last values, and whether the
@@ -3019,25 +2697,6 @@
       $deny_frozen_access(this);
       return this[ivar] = static_val;
     }
-  }
-
-  // Primitives for handling parameters
-  Opal.ensure_kwargs = function(kwargs) {
-    if (kwargs == null) {
-      return new Map();
-    } else if (kwargs.$$is_hash) {
-      return kwargs;
-    } else {
-      $raise(Opal.ArgumentError, 'expected kwargs');
-    }
-  }
-
-  Opal.get_kwarg = function(kwargs, key) {
-    var kwarg = Opal.hash_get(kwargs, key);
-    if (kwarg === undefined) {
-      $raise(Opal.ArgumentError, 'missing keyword: '+key);
-    }
-    return kwarg;
   }
 
   // Arrays of size > 32 elements that contain only strings,
