+++ opal/runtime/runtime.rb
@@ -1,2 +1,3 @@
-require "runtime/boot"
-require "runtime/helpers"
+require 'runtime/boot'
+require 'runtime/hash'
+require 'runtime/helpers'
